Add support for username and password in the PDO DSN
This implements a new feature where PDO supports user and password being
specified in the DSN instead of separate constructor params. This is
backwards compatible and it ignores the credentials in DSN when separate
params are passed.

This additionally allows a security feature where user-space is unable
to read the credentials but PDO is able to connect anyway.